+++ src/Network/Unexceptional/Bytes.hs
@@ -6,26 +6,35 @@
 
 module Network.Unexceptional.Bytes
   ( send
+  , sendInterruptible
   , receive
+  , receiveInterruptible
   ) where
 
+import Control.Applicative ((<|>))
+import Control.Concurrent.STM (STM,TVar)
 import Control.Exception (throwIO)
+import Control.Monad ((<=<))
 import Control.Monad (when)
 import Data.Array.Byte (ByteArray)
 import Data.Bytes.Types (Bytes(Bytes),MutableBytes(MutableBytes))
+import Data.Functor (($>))
+import Data.Primitive (MutableByteArray)
 import Foreign.C.Error (Errno)
 import Foreign.C.Error.Pattern (pattern EWOULDBLOCK,pattern EAGAIN)
-import GHC.Conc (threadWaitWrite)
+import GHC.Conc (threadWaitWrite,threadWaitWriteSTM)
+import GHC.Exts (RealWorld)
 import Network.Socket (Socket)
 import System.Posix.Types (Fd(Fd))
 
+import qualified Control.Concurrent.STM as STM
+import qualified Data.Bytes.Types
+import qualified Data.Primitive as PM
+import qualified Linux.Socket as X
+import qualified Network.Socket as S
+import qualified Network.Unexceptional.MutableBytes as MB
 import qualified Network.Unexceptional.Types as Types
 import qualified Posix.Socket as X
-import qualified Linux.Socket as X
-import qualified Data.Bytes.Types
-import qualified Network.Socket as S
-import qualified Data.Primitive as PM
-import qualified Network.Unexceptional.MutableBytes as MB
 
 -- | Send the entire byte sequence. This call POSIX @send@ in a loop
 -- until all of the bytes have been sent.
@@ -60,6 +69,49 @@ sendLoop !fd !arr !off !len =
             LT -> sendLoop fd arr (off + sentSz) (len - sentSz)
             GT -> fail "Network.Unexceptional.Bytes.sendAll: send claimed to send too many bytes"
 
+sendInterruptibleLoop :: TVar Bool -> Fd -> ByteArray -> Int -> Int -> IO (Either Errno ())
+sendInterruptibleLoop !interrupt !fd !arr !off !len =
+  X.uninterruptibleSendByteArray fd arr off (fromIntegral len) (X.noSignal <> X.dontWait) >>= \case
+    Left e -> if e == EAGAIN || e == EWOULDBLOCK
+      then waitUntilWriteable interrupt fd >>= \case
+        Ready -> sendInterruptibleLoop interrupt fd arr off len
+        Interrupted -> pure (Left EAGAIN)
+      else pure (Left e)
+    Right sentSzC ->
+      let sentSz = fromIntegral sentSzC :: Int
+       in case compare sentSz len of
+            EQ -> pure (Right ())
+            LT -> sendInterruptibleLoop interrupt fd arr (off + sentSz) (len - sentSz)
+            GT -> fail "Network.Unexceptional.Bytes.sendAll: send claimed to send too many bytes"
+
+-- | Variant of 'send' that fails with @EAGAIN@ if the interrupt ever becomes true.
+sendInterruptible ::
+     TVar Bool
+  -> Socket
+  -> Bytes
+  -> IO (Either Errno ())
+sendInterruptible !interrupt s Bytes{array,offset,length=len} = case len of
+  0 -> pure (Right ())
+  _ -> STM.readTVarIO interrupt >>= \case
+    True -> pure (Left EAGAIN)
+    False -> S.withFdSocket s $ \fd ->
+      -- We attempt the first send without testing if the socket is in
+      -- ready for writes. This is because it is uncommon for the transmit
+      -- buffer to already be full.
+      sendInterruptibleLoop interrupt (Fd fd) array offset len
+
+data Outcome = Ready | Interrupted
+
+checkFinished :: TVar Bool -> STM ()
+checkFinished = STM.check <=< STM.readTVar
+
+waitUntilWriteable :: TVar Bool -> Fd -> IO Outcome
+waitUntilWriteable !interrupt !fd = do
+  (isReadyAction,deregister) <- threadWaitWriteSTM fd
+  outcome <- STM.atomically $ (isReadyAction $> Ready) <|> (checkFinished interrupt $> Interrupted)
+  deregister
+  pure outcome
+
 -- | If this returns zero bytes, it means that the peer has
 -- performed an orderly shutdown.
 receive :: 
@@ -69,12 +121,26 @@ receive ::
 receive s n = if n > 0
   then do
     dst <- PM.newByteArray n
-    MB.receive s (MutableBytes dst 0 n) >>= \case
-      Left e -> pure (Left e)
-      Right m -> do
-        when (m < n) (PM.shrinkMutableByteArray dst m)
-        dst' <- PM.unsafeFreezeByteArray dst 
-        pure (Right (Bytes dst' 0 m))
+    MB.receive s (MutableBytes dst 0 n) >>= handleRececeptionResult dst n
   else throwIO Types.NonpositiveReceptionSize
 
+-- | If this returns zero bytes, it means that the peer has
+-- performed an orderly shutdown.
+receiveInterruptible :: 
+     TVar Bool -- ^ Interrupt
+  -> Socket
+  -> Int -- ^ Maximum number of bytes to receive
+  -> IO (Either Errno Bytes)
+receiveInterruptible !interrupt s n = if n > 0
+  then do
+    dst <- PM.newByteArray n
+    MB.receiveInterruptible interrupt s (MutableBytes dst 0 n) >>= handleRececeptionResult dst n
+  else throwIO Types.NonpositiveReceptionSize
 
+handleRececeptionResult :: MutableByteArray RealWorld -> Int -> Either Errno Int -> IO (Either Errno Bytes)
+handleRececeptionResult !dst !n x = case x of
+  Left e -> pure (Left e)
+  Right m -> do
+    when (m < n) (PM.shrinkMutableByteArray dst m)
+    dst' <- PM.unsafeFreezeByteArray dst 
+    pure (Right (Bytes dst' 0 m))