+++ src/Network/Unexceptional/Chunks.hs
@@ -6,11 +6,13 @@
 
 module Network.Unexceptional.Chunks
   ( send
+  , sendInterruptible
   ) where
 
 import Foreign.C.Error (Errno)
 import Network.Socket (Socket)
 import Data.Bytes.Chunks (Chunks)
+import Control.Concurrent.STM (TVar)
 
 import qualified Data.Bytes.Chunks as Chunks
 import qualified Network.Unexceptional.Bytes as NB
@@ -22,3 +24,10 @@ send ::
   -> Chunks
   -> IO (Either Errno ())
 send !s cs = NB.send s (Chunks.concat cs)
+
+sendInterruptible ::
+     TVar Bool
+  -> Socket
+  -> Chunks
+  -> IO (Either Errno ())
+sendInterruptible !interrupt !s cs = NB.sendInterruptible interrupt s (Chunks.concat cs)