+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Revision history for network-unexceptional
 
+## 0.1.3.0 -- 2023-08-30
+
+* Add interruptible send and receive functions to give users control over
+  when to give to abandon communicating.
+* Add these functions to `Network.Unexceptional`: `connect`,
+  `connectInterruptible`, `socket`.
+
 ## 0.1.1.0 -- 2023-08-14
 
 * Add Network.Unexceptional module with `accept_`.